requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
kubernetes==29.0.0
prometheus-client==0.19.0
openai>=1.12.0
python-dotenv==1.0.0
